Implement simplified ambient occlusion
Ambient occlusion is implemented by tracing randomly
generated rays towards the hemisphere for each intersection.
If the ray intersects a primitive within a certain distance
then the point is considered occluded for that ray. No
weighing based on distance is done in this simplified model,
i.e. all intersections within the distance threshold are
considered to cause an equal amount of occlusion.